rag_chatbot_korean/requirements.txt
streamlit==1.38.0
langchain==0.2.16
langchain-community==0.2.8
langchain-openai==0.1.23
faiss-cpu==1.8.0.post1
pypdf==4.3.1
python-dotenv==1.0.1
tiktoken==0.7.0
